+++ models/tx.go
@@ -1,0 +1,10 @@
+package models
+
+const (
+	TxBasic = "basic"
+	TxOrder = "order"
+)
+
+type Tx interface {
+	Type() string
+}
